Added doc comments to controller

+++ src/main/java/com/dezoo/personeel/controllers/PersonnelController.java
@@ -18,24 +18,45 @@ public class PersonnelController {
     @Autowired
     private PersonelRepository personelRepository;
 
+    /**
+     * Gets all personnel from the db
+     * @return A list of PersonnelMember objects, representing all of the personnel of the zoo
+     */
     @GetMapping
     public List<PersonnelMember> getPersonnelMember() {
         return personelRepository.findAll();
     }
 
+    /**
+     * Gets a single personnel member from the db
+     * @param personnelID   The personnelID of the personnel member that has to be returned
+     * @return              The personnel member which corresponds with the given personnelID
+     */
     @GetMapping("/{personnelID}")
     public PersonnelMember getPersonnelMemberID(@PathVariable String personnelID) {
         return personelRepository.findPersonelMemberByPersonelId(personnelID);
     }
 
+    /**
+     * Adds a personnel member to the db
+     * @param personnelMember   The personnel member that has to be added
+     * @return                  The newly added personnel member
+     */
     @PostMapping
-    public void postPersonnelMember(@RequestBody PersonnelMember personnelMember) {
+    public ResponseEntity<PersonnelMember> postPersonnelMember(@RequestBody PersonnelMember personnelMember) {
         personelRepository.save(personnelMember);
+
+        return ResponseEntity.ok(personnelMember);
     }
 
-    @PutMapping("/{personnelID}")
-    public ResponseEntity<PersonnelMember> putPersonnelMember(@PathVariable String personnelID, @RequestBody PersonnelMember personnelMember) {
-        PersonnelMember personnelMember2 = personelRepository.findPersonelMemberByPersonelId(personnelID);
+    /**
+     * Updates a personnel member
+     * @param personnelMember   The personnel member that has to be updated
+     * @return                  The newly updated personnel member
+     */
+    @PutMapping
+    public ResponseEntity<PersonnelMember> putPersonnelMember(@RequestBody PersonnelMember personnelMember) {
+        PersonnelMember personnelMember2 = personelRepository.findPersonelMemberByPersonelId(personnelMember.getPersonelId());
 
         personnelMember2.setAddress(personnelMember.getAddress());
         personnelMember2.setDateOfBirth(personnelMember.getDateOfBirth());
@@ -45,17 +66,24 @@ public class PersonnelController {
         personnelMember2.setPrivatePhoneNumber(personnelMember.getPrivatePhoneNumber());
         personnelMember2.setPersonelCategory(personnelMember.getPersonelCategory());
 
-        personnelMember2 = personelRepository.save(personnelMember2);
+        personelRepository.save(personnelMember2);
 
         return ResponseEntity.ok(personnelMember2);
     }
 
+    /**
+     * Deletes a personnel member
+     * @param personnelID   The personnelID of the personnel member that has to be deleted
+     * @return              A response with either a "Not Found" code or an "OK" code
+     */
     @DeleteMapping("/{personnelID}")
-    public ResponseEntity<PersonnelMember> deletePersonnelMember(@PathVariable String personnelID){
+    public ResponseEntity<PersonnelMember> deletePersonnelMember(@PathVariable String personnelID) {
         PersonnelMember personnelMember = personelRepository.findPersonelMemberByPersonelId(personnelID);
-        personelRepository.delete(personnelMember);
-
-        return ResponseEntity.ok(personnelMember);
+        if (personnelMember != null) {
+            personelRepository.delete(personnelMember);
+            return ResponseEntity.ok().build();
+        } else
+            return ResponseEntity.notFound().build();
     }
 
     @PostConstruct
@@ -64,7 +92,7 @@ public class PersonnelController {
         try {
             personelRepository.save(new PersonnelMember("fs161100", "Ferre", "Snyers", format.parse("16/11/2000"), "Gestelstraat 21", "2250", "+32441439", "Administration"));
             personelRepository.save(new PersonnelMember("cn170999", "Christophe", "Neefs", format.parse("17/09/1999"), "Lier", "2500", "", "Rabbits"));
-            personelRepository.save(new PersonnelMember("rh031000", "Robbe", "Heremans", format.parse("03/10/2000"), "Westerloo", "2260", "", "Lions"));
+            personelRepository.save(new PersonnelMember("rh031000", "Robbe", "Heremans", format.parse("03/10/2000"), "Westerlo", "2260", "", "Lions"));
         } catch (ParseException ignored) {
         }
     }
